+++ Changelog.md
@@ -3,6 +3,7 @@
 - Reverted `_AnyCollection_` to the less specific version and settled it not being a true Any but instead a "these are all types that TFWR can support match one of them"
 - Made `list` `V` more specific as AnyTFWR
 - Made `dict` `K` more general as Any
+- Made `Drone` `R` more general as Any
 
 # May/24 Update 1